# frontend/src/integrations/api/api-spec.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"description":"This operation creates both a user and a supabase auth account","operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserWithAccountDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"content":{"application/json":{"examples":{"BadRequestException":{"description":"Bad Request","value":{"statusCode":400,"message":"Bad Request","error":"Bad Request"}}},"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"Bad Request"},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message"]}}},"description":""},"500":{"content":{"application/json":{"examples":{"InternalServerErrorException":{"description":"Internal Server Error","value":{"statusCode":500,"message":"Internal Server Error","error":"Internal Server Error"}}},"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":500},"message":{"type":"string","example":"Internal Server Error"},"error":{"type":"string","example":"Internal Server Error"}},"required":["statusCode","message"]}}},"description":""}},"security":[{"bearer":[]}],"summary":"Create a new user","tags":["Users"]},"get":{"operationId":"UsersController_findAll","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"role","required":false,"in":"query","schema":{"type":"string","enum":["student","mentor","admin"]}}],"responses":{"200":{"description":"List of users retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedUsersDto"}}}},"400":{"content":{"application/json":{"examples":{"BadRequestException":{"description":"Bad Request","value":{"statusCode":400,"message":"Bad Request","error":"Bad Request"}}},"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"Bad Request"},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message"]}}},"description":""},"500":{"content":{"application/json":{"examples":{"InternalServerErrorException":{"description":"Internal Server Error","value":{"statusCode":500,"message":"Internal Server Error","error":"Internal Server Error"}}},"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":500},"message":{"type":"string","example":"Internal Server Error"},"error":{"type":"string","example":"Internal Server Error"}},"required":["statusCode","message"]}}},"description":""}},"security":[{"bearer":[]}],"summary":"Retrieves a paginated list of users based on the provided filter parameters.\n\n- **Access:** Requires `ADMIN` role.\n- **Filtering & Pagination:** Uses the `FilterUserDto` to define query parameters such as search terms, sorting, and page size.","tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[],"responses":{"200":{"description":"User found successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"content":{"application/json":{"examples":{"BadRequestException":{"description":"Bad Request","value":{"statusCode":400,"message":"Bad Request","error":"Bad Request"}}},"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"Bad Request"},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message"]}}},"description":""},"404":{"content":{"application/json":{"examples":{"NotFoundException":{"description":"Not Found","value":{"statusCode":404,"message":"Not Found","error":"Not Found"}}},"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":404},"message":{"type":"string","example":"Not Found"},"error":{"type":"string","example":"Not Found"}},"required":["statusCode","message"]}}},"description":""},"500":{"content":{"application/json":{"examples":{"InternalServerErrorException":{"description":"Internal Server Error","value":{"statusCode":500,"message":"Internal Server Error","error":"Internal Server Error"}}},"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":500},"message":{"type":"string","example":"Internal Server Error"},"error":{"type":"string","example":"Internal Server Error"}},"required":["statusCode","message"]}}},"description":""}},"security":[{"bearer":[]}],"summary":"Retrieves a specific user by their unique identifier.\n\n- **Validation:** Ensures the provided `id` is a valid identifier format.\n- **Not Found Handling:** Throws an error if no matching user is found.","tags":["Users"]}},"/courses":{"post":{"operationId":"CoursesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCourseDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Courses"]},"get":{"operationId":"CoursesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Courses"]}},"/courses/{id}":{"get":{"operationId":"CoursesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Courses"]},"patch":{"operationId":"CoursesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCourseDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Courses"]},"delete":{"operationId":"CoursesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Courses"]}},"/test/student":{"get":{"operationId":"TestController_testStudent","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Test"]}},"/test/admin":{"get":{"operationId":"TestController_testAdmin","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Test"]}}},"info":{"title":"API","description":"Description","version":"1.0","contact":{}},"tags":[{"name":"api","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Role":{"type":"string","enum":["student","mentor","admin"]},"UserCredentialsDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["email","password"]},"CreateUserWithAccountDto":{"type":"object","properties":{"firstName":{"type":"string"},"middleName":{"type":"string","nullable":true},"lastName":{"type":"string"},"role":{"allOf":[{"$ref":"#/components/schemas/Role"}]},"credentials":{"$ref":"#/components/schemas/UserCredentialsDto"}},"required":["firstName","lastName","role"]},"User":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string","nullable":true},"lastName":{"type":"string"},"role":{"allOf":[{"$ref":"#/components/schemas/Role"}]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","nullable":true,"format":"date-time"}},"required":["id","firstName","middleName","lastName","role","createdAt","updatedAt","deletedAt"]},"UserAccount":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"authUid":{"type":"string"},"email":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","userId","authUid","createdAt","updatedAt"]},"UserDetails":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"dob":{"format":"date-time","type":"string"},"gender":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","userId","createdAt","updatedAt"]},"UserWithRelations":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string"},"lastName":{"type":"string"},"role":{"allOf":[{"$ref":"#/components/schemas/Role"}]},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"},"userAccount":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/UserAccount"}]},"userDetails":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/UserDetails"}]}},"required":["id","firstName","lastName","role","createdAt","updatedAt","userAccount","userDetails"]},"PaginationMetaDto":{"type":"object","properties":{"isFirstPage":{"type":"boolean","example":true},"isLastPage":{"type":"boolean","example":true},"currentPage":{"type":"number","example":1},"previousPage":{"type":"number","nullable":true,"example":null},"nextPage":{"type":"number","nullable":true,"example":null},"pageCount":{"type":"number","example":1},"totalCount":{"type":"number","example":42}},"required":["isFirstPage","isLastPage","currentPage","previousPage","nextPage","pageCount","totalCount"]},"PaginatedUsersDto":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/UserWithRelations"}},"meta":{"additionalProperties":true,"nullable":false,"allOf":[{"$ref":"#/components/schemas/PaginationMetaDto"}]}},"required":["users","meta"]},"CreateCourseDto":{"type":"object","properties":{}},"UpdateCourseDto":{"type":"object","properties":{}}}}}